+++ napari_animation/animation.py
@@ -1,138 +1,13 @@
 import os
-from dataclasses import asdict
 from itertools import count
 from pathlib import Path
-from typing import Dict, Iterator, Sequence, Tuple
 
 import imageio
-import numpy as np
-from napari.utils.events import EmitterGroup, SelectableEventedList
 from napari.utils.io import imsave
-from scipy import ndimage as ndi
 
 from .easing import Easing
-from .interpolation import Interpolation
-from .key_frame import KeyFrame, ViewerState
-from .utils import pairwise
-
-
-class AnimationGenerator(Sequence[ViewerState]):
-    def __init__(self, key_frames: SelectableEventedList[KeyFrame]) -> None:
-        super().__init__()
-        self.key_frames = key_frames
-        key_frames.events.inserted.connect(self._update_keyframe_map)
-        key_frames.events.removed.connect(self._update_keyframe_map)
-        key_frames.events.changed.connect(self._update_keyframe_map)
-        key_frames.events.reordered.connect(self._update_keyframe_map)
-
-        self.events = EmitterGroup(source=self, n_frames=None)
-
-        self.state_interpolation_map = {
-            "camera.angles": Interpolation.SLERP,
-            "camera.zoom": Interpolation.LOG,
-        }
-
-        # cache of interpolated viewer states
-        self._cache: Dict[int, ViewerState] = {}
-        # map of frame number -> (kf0, kf1, fraction)
-        self._keyframe_map: Dict[int, Tuple[KeyFrame, KeyFrame, float]] = {}
-        self._update_keyframe_map()
-
-    def _update_keyframe_map(self, event=None):
-        """Create a map of frame number -> (kf0, kf1, fraction)"""
-        self._keyframe_map.clear()
-        self._cache.clear()
-        if len(self.key_frames) < 2:
-            self.events.n_frames(value=len(self))
-            return
-
-        f = 0
-        for kf0, kf1 in pairwise(self.key_frames):
-            for s in range(kf1.steps):
-                fraction = s / kf1.steps
-                self._keyframe_map[f] = (kf0, kf1, fraction)
-                f += 1
-        self._keyframe_map[f] = (kf1, kf1, 0)
-        self.events.n_frames(value=len(self))
-
-    def __len__(self) -> int:
-        """The total frame count of the animation"""
-        return len(self._keyframe_map)
-
-    def __getitem__(self, key: int) -> ViewerState:
-        """Get the interpolated state at frame `key` in the animation."""
-        if key not in self._cache:
-            kf0, kf1, frac = self._keyframe_map[key]
-            if frac == 0:
-                self._cache[key] = kf0.viewer_state
-            else:
-                self._cache[key] = self._interpolate_state(
-                    kf0.viewer_state,
-                    kf1.viewer_state,
-                    frac,
-                    self.state_interpolation_map,
-                )
-
-        return self._cache[key]
-
-    def _interpolate_state(
-        self,
-        from_state: ViewerState,
-        to_state: ViewerState,
-        fraction,
-        state_interpolation_map=None,
-    ):
-        """Interpolate a state between two states
-
-        Parameters
-        ----------
-        initial_state : dict
-            Description of initial viewer state.
-        final_state : dict
-            Description of final viewer state.
-        fraction : float
-            Interpolation fraction, must be between `0` and `1`.
-            A value of `0` will return the initial state. A
-            value of `1` will return the final state.
-        state_interpolation_map : dict
-            Dictionary relating state attributes to interpolation functions.
-
-        Returns
-        -------
-        state : dict
-            Description of viewer state.
-        """
-        from .utils import keys_to_list, nested_get, nested_set
-
-        if state_interpolation_map is None:
-            state_interpolation_map = self.state_interpolation_map
-
-        state = {}
-        separator = "."
-
-        from_state = asdict(from_state)
-        to_state = asdict(to_state)
-
-        for keys in keys_to_list(from_state):
-            v0 = nested_get(from_state, keys)
-            v1 = nested_get(to_state, keys)
-
-            property_string = separator.join(keys)
-
-            if property_string in state_interpolation_map.keys():
-                interpolation_func = state_interpolation_map[property_string]
-            else:
-                interpolation_func = Interpolation.DEFAULT
-
-            nested_set(state, keys, interpolation_func(v0, v1, fraction))
-
-        return ViewerState(**state)
-
-    def iter_frames(self, viewer, canvas_only=True) -> Iterator[np.ndarray]:
-        n_frames = len(self)
-        for i, state in enumerate(self):
-            print("Rendering frame ", i + 1, "of", n_frames)
-            yield state.render(viewer, canvas_only=canvas_only)
+from .frame_sequence import FrameSequence
+from .key_frame import KeyFrame, KeyFrameList
 
 
 class Animation:
@@ -147,23 +22,18 @@ class Animation:
     ----------
     key_frames : list of dict
         List of viewer state dictionaries.
-    frame : int
-        Currently shown key frame.
-    state_interpolation_map : dict
-        Dictionary relating state attributes to interpolation functions.
     """
 
     def __init__(self, viewer):
         self.viewer = viewer
 
-        self.key_frames: SelectableEventedList[
-            KeyFrame
-        ] = SelectableEventedList(basetype=KeyFrame)
+        self.key_frames = KeyFrameList()
         self.key_frames.selection.events._current.connect(
             self._on_current_keyframe_changed
         )
-        self._generator = AnimationGenerator(self.key_frames)
         self._keyframe_counter = count()  # track number of frames created
+
+        self._frames = FrameSequence(self.key_frames)
 
     def capture_keyframe(
         self, steps=15, ease=Easing.LINEAR, insert=True, position: int = None
@@ -304,12 +174,13 @@ class Animation:
                 folder_path.mkdir(exist_ok=True)
 
         # create a frame generator
-        frames = self._generator.iter_frames(self.viewer, canvas_only)
+        frames = self._frames.iter_frames(
+            self.viewer, canvas_only, scale_factor
+        )
+        n_frames = len(self._frames)
         # save frames
         for ind, frame in enumerate(frames):
-            if scale_factor is not None:
-                frame = ndi.zoom(frame, (scale_factor, scale_factor, 1))
-                frame = frame.astype(np.uint8)
+            print("Rendering frame ", ind + 1, "of", n_frames)
             if not save_as_folder:
                 writer.append_data(frame)
             else:
@@ -326,10 +197,10 @@ class Animation:
     def set_movie_frame(self, frame: int):
         """Set state to a specific frame in the final movie."""
         try:
-            self._generator[frame].apply_to_viewer(self.viewer)
+            self._frames[frame].apply_to_viewer(self.viewer)
         except KeyError:
             return
 
         with self.key_frames.selection.events._current.blocker():
-            frame = self._generator._keyframe_map[frame][0]
+            frame = self._frames._frame_index[frame][0]
             self.key_frames.selection.active = frame
